+++ src/index.ts
@@ -1,6 +1,11 @@
 // Copyright (c) 2020 Antti Kivi
 // Licensed under the MIT License
 
-import {run} from "./main";
+import {run, upload} from "./main";
+import * as stateHelper from "./state-helper";
 
-run();
+if (!stateHelper.IS_POST) {
+  run();
+} else {
+  upload();
+}