+++ src/state-helper.ts
@@ -1,0 +1,7 @@
+// Copyright (c) 2020 Antti Kivi
+// Licensed under the MIT License
+
+/**
+ * Indicates whether the POST action is running
+ */
+export const IS_POST = !!process.env["STATE_isPost"];